allow consumer to keep running when published messages are exhausted
Previously, a consumer task would simply end if there are no more
messages on the queue being consumed.  This is counter to the
expectation that a consumer will keep running and consume new messages
that appear on the queue.

For simple tests this expectation can be ignored by pre-loading the
queue with a few messages, seeing if they're consumed, then exit the
test.  But for testing more complicated interactions where a consumer
might need to wait for messages from a publisher this won't work.

I believe this maintains the simplicity of previous use cases, except
that Storage.get_next_message() is never exhausted, and can block.

To this end it's also necessary (when it wasn't previously) to ensure
that all consumer tasks associated with a connection are cancelled when
closing that connection.